+++ tests/insertion.rs
@@ -27,7 +27,8 @@ fn table_storage() {
                 ..Default::default()
             }),
         ))
-        .replicate::<TableComponent>();
+        .replicate::<TableComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -64,7 +65,8 @@ fn sparse_set_storage() {
                 ..Default::default()
             }),
         ))
-        .replicate::<SparseSetComponent>();
+        .replicate::<SparseSetComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -101,7 +103,8 @@ fn immutable() {
                 ..Default::default()
             }),
         ))
-        .replicate::<ImmutableComponent>();
+        .replicate::<ImmutableComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -151,7 +154,8 @@ fn mapped_existing_entity() {
                 ..Default::default()
             }),
         ))
-        .replicate::<MappedComponent>();
+        .replicate::<MappedComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -200,7 +204,8 @@ fn mapped_new_entity() {
                 ..Default::default()
             }),
         ))
-        .replicate::<MappedComponent>();
+        .replicate::<MappedComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -246,7 +251,8 @@ fn multiple_components() {
             }),
         ))
         .replicate::<ComponentA>()
-        .replicate::<ComponentB>();
+        .replicate::<ComponentB>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -291,7 +297,8 @@ fn command_fns() {
             }),
         ))
         .replicate::<OriginalComponent>()
-        .set_command_fns(replace, command_fns::default_remove::<ReplacedComponent>);
+        .set_command_fns(replace, command_fns::default_remove::<ReplacedComponent>)
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -335,7 +342,8 @@ fn marker() {
         .set_marker_fns::<ReplaceMarker, _>(
             replace,
             command_fns::default_remove::<ReplacedComponent>,
-        );
+        )
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -382,7 +390,8 @@ fn group() {
                 ..Default::default()
             }),
         ))
-        .replicate_bundle::<(ComponentA, ComponentB)>();
+        .replicate_bundle::<(ComponentA, ComponentB)>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -418,7 +427,8 @@ fn not_replicated() {
                 tick_policy: TickPolicy::EveryFrame,
                 ..Default::default()
             }),
-        ));
+        ))
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -455,7 +465,8 @@ fn after_removal() {
                 ..Default::default()
             }),
         ))
-        .replicate::<TestComponent>();
+        .replicate::<TestComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -507,7 +518,8 @@ fn before_started_replication() {
                 ..Default::default()
             }),
         ))
-        .replicate::<TestComponent>();
+        .replicate::<TestComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -553,7 +565,8 @@ fn after_started_replication() {
                 ..Default::default()
             }),
         ))
-        .replicate::<TestComponent>();
+        .replicate::<TestComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);
@@ -592,7 +605,8 @@ fn confirm_history() {
                 ..Default::default()
             }),
         ))
-        .replicate::<TestComponent>();
+        .replicate::<TestComponent>()
+        .finish();
     }
 
     server_app.connect_client(&mut client_app);